requirements.txt
streamlit
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
